yt-live/pyproject.toml
[tool.poetry]
name = "yt-live"
version = "0.1.0"
description = ""
authors = ["Keisuke Hirota <tahiro.k.ad@gmail.com>"]
readme = "README.md"
packages = [{include = "yt_live"}]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
ffmpeg-python = "^0.2.0"
typer = "^0.7.0"

[tool.poetry.scripts]
yt_live = "yt_live.__main__:main"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
